requirements.txt
numpy 
lalsuite
h5py
scipy 
sxs 
requests_cache
